Re-add old contract index files
