+++ src/contracts/evm/Contract.ts
@@ -7,7 +7,7 @@ import {
 } from "types/evm";
 import { WalletState } from "contexts/WalletContext";
 import { toTuple } from "helpers";
-import Account from "./Account";
+import Account from "./Account/Account";
 
 /** Represents generic contract functionality. Needs to be extended by actual contracts like {@link Account}. */
 export default class Contract {
